Tree shake and bundle with esbuild
